Mixed the apis
use this to be able to search with many stores:
http://127.0.0.1:8000/api/start/?keyword=laptop&storeList=Ebay,Amazon,AliExpress

Currently only these 3 work, Shein works but has a loop issue yet to be fixed.

The store images will be returned as links of file location (Daber 7alak)

Amazon api needs to return catagories

Descreption of items yet to be implemented.